Add various algorithms for data structures and optimization problems
- Implemented Divide and Conquer approach for finding minimum and maximum in an array.
- Added Merge Sort algorithm for sorting an array.
- Implemented Prim's algorithm for finding the Minimum Spanning Tree (MST) in a graph.
- Added Quick Sort algorithm for efficient sorting.
- Implemented Traveling Salesman Problem (TSP) using both Dynamic Programming and Branch and Bound methods.
- Added 0/1 Knapsack problem solution using brute force.
- Implemented Fractional Knapsack problem using a greedy approach.
- Added Prim's algorithm using adjacency list representation.
- Developed a campus network design optimizer with functionalities for BFS, Kruskal's algorithm, Dijkstra's algorithm, and Floyd-Warshall algorithm.
- Removed unused DFS and merge sort implementations.

+++ DAA Lab/knapSack_01_Brute.c
@@ -6,11 +6,11 @@ typedef struct Node {
     int weight;
 } node;
 
-float knapSack(node* arr, int n, int cap);
+int knapSack(node* arr, int n, int cap);
 
 int main() {
     int n, cap;
-    float max;
+    int max;
 
     printf("Enter the number of items: ");
     scanf("%d", &n);
@@ -27,14 +27,14 @@ int main() {
 
     max = knapSack(items, n, cap);
 
-    printf("Maximum Profit Value: %.2f\n", max);
+    printf("Maximum Profit Value: %d\n", max);
 
     free(items);
     return 0;
 }
 
-float knapSack(node* arr, int n, int cap) {
-    float maxProfit = 0;
+int knapSack(node* arr, int n, int cap) {
+    int maxProfit = 0;
 
     int totalCombo = 1;
     for (int i = 0; i < n; i++) {
